refactor: Pipeline 业务逻辑剥离 + 协议对齐
P0 架构修复:
- Pipeline: 删除 hardcoded think/execute 阶段消息组装代码
- SlotAccessPoint: 新增 append_message() 方法 + messages:append 权限
- StepContext: 实现 append_message()
- 新建 ThoughtSyncSlot (Phase::think 末尾)
- 新建 ObservationSyncSlot (Phase::execute 末尾)
- main.rs 注册两个新 Slot

P1 协议违规修复:
- K-R01: llm_thinker_slot provider_raw(session-context) 替换为 PROVIDER_SESSION_CONTEXT
- clippy: 修复 collapsible_match + unnecessary_cast

P3 文档同步:
- protocol-Slot接入协议.md: init() 调用方 / append_message
- protocol-Service集成协议.md: provider_raw / unregister_provider
- AI开发红线与纪律.md: 更新已知偏差表
- llm_thinker_design.md: messages.push() 替换为 ThoughtSyncSlot 流程

验证: cargo fmt --check + clippy --lib -D warnings + test --lib (315/315)

+++ Agnet-rust/src/plugins/services/llm/retry.rs
@@ -24,14 +24,13 @@ impl RetryManager {
 
     /// 添加 ±25% 随机 jitter 避免 thundering herd
     fn add_jitter(delay: Duration, attempt: u32) -> Duration {
-        let nanos = delay.as_nanos() as u128;
+        let nanos = delay.as_nanos();
         if nanos == 0 {
             return delay;
         }
         // 基于 attempt 和 nanos 的简单伪随机：0..=50 百分比
         let seed = (nanos.wrapping_mul(1103515245).wrapping_add(12345)
-            ^ (attempt as u128).wrapping_mul(6364136223846793005))
-            as u64;
+            ^ (attempt as u128).wrapping_mul(6364136223846793005)) as u64;
         let pct = (seed % 51) as f64; // 0..50
         let jitter_factor = 0.75 + (pct / 100.0); // 0.75..1.25
         Duration::from_nanos((nanos as f64 * jitter_factor) as u64)
